@@ -71,67 +123,75 @@
 	"author": "Bitfocus AS",
 	"license": "MIT",
 	"devDependencies": {
+		"@babel/cli": "^7.12.1",
+		"@babel/core": "^7.12.3",
+		"@babel/preset-env": "^7.12.1",
+		"asar": "^3.0.3",
 		"chai": "^4.2.0",
 		"chai-fs": "^2.0.0",
-		"electron": "^2.0.8",
-		"electron-builder": "20.10.0",
-		"mocha": "^6.1.4",
-		"nodemon": "^2.0.2",
-		"s3": "^4.4.0",
+		"electron": "^10.3.1",
+		"electron-builder": "^22.9.1",
+		"mocha": "^6.2.3",
+		"nodemon": "^2.0.6",
+		"p-all": "^3.0.0",
+		"rimraf": "^3.0.2",
 		"s3-upload-stream": "^1.0.7"
 	},
 	"dependencies": {
-		"@coreui/coreui": "^2.0.0-rc.2",
-		"@coreui/coreui-plugin-chartjs-custom-tooltips": "1.2.0",
+		"@coreui/coreui": "^2.1.16",
+		"@sentry/electron": "^1.3.0",
 		"app-root-path": "^2.0.1",
-		"aws-sdk": "^2.656.0",
+		"aws-sdk": "^2.783.0",
 		"binopsy": "^0.0.0",
-		"bootstrap": "^4.2.1",
+		"bootstrap": "^4.5.3",
 		"bootstrap-colorpicker": "^2.5.2",
-		"chart.js": "^2.7.2",
 		"check-ip": "^1.1.1",
-		"debug": "^4.1.1",
+		"debug": "^4.2.0",
 		"electron-debug": "^1.5.0",
 		"electron-rpc": "^2.0.1",
-		"elgato-stream-deck": "^3.3.0",
+		"elgato-stream-deck": "^4.0.0",
+		"emberplus-connection": "^0.0.4",
 		"express": "^4.16.3",
-		"find-process": "^1.4.3",
-		"flag-icon-css": "3.4.6",
+		"find-process": "1.4.3",
+		"flag-icon-css": "3.5.0",
 		"font-awesome": "4.7.0",
-		"fs-extra": "^7.0.0",
+		"fs-extra": "^9.0.1",
 		"infinitton-idisplay": "^1.0.5",
-		"jquery": ">=3.4.0",
-		"json-parser": "^1.1.5",
-		"lodash": "^4.17.10",
+		"jquery": ">=3.5.1",
+		"lodash": "^4.17.20",
 		"marked": "0.7.0",
-		"mkdirp": "^0.5.1",
-		"moment": "^2.22.2",
+		"mkdirp": "^0.5.5",
+		"moment": "^2.29.1",
 		"network": "^0.4.1",
+		"node-fetch": "^2.6.1",
 		"node-rest-client": "^3.1.0",
-		"opencollective": "^1.0.3",
-		"opencollective-postinstall": "^2.0.1",
-		"osc": "^2.2.2",
+		"osc": "^2.4.1",
 		"pace-progress": "1.0.2",
-		"perfect-scrollbar": "^1.3.0",
+		"perfect-scrollbar": "^1.5.0",
 		"ping": "^0.2.2",
 		"pngjs": "^3.3.3",
 		"popper.js": "1.16.1",
-		"read-package-json": "^2.0.13",
+		"read-package-json": "^2.1.2",
 		"select2": "4.0.13",
-		"sharp": "^0.23.3",
-		"shortid": "^2.2.8",
-		"simple-line-icons": "2.4.1",
-		"socket.io": "^2.1.0",
-		"socket.io-client": "^2.1.0",
+		"sharp": "^0.27.1",
+		"shortid": "^2.2.16",
+		"socket.io": "^2.3.0",
+		"socket.io-client": "^2.3.1",
 		"strip-ansi": "^5.2.0",
 		"tributejs": "github:peschuster/tribute",
-		"websocket": "^1.0.28"
+		"websocket": "^1.0.32",
+		"xkeys": "^1.1.1"
 	},
 	"collective": {
 		"type": "donorbox",
 		"url": "https://donorbox.org/bitfocus-opensource"
 	},
 	"optionalDependencies": {
-		"@julusian/jpeg-turbo": "^0.5.4"
+		"@julusian/jpeg-turbo": "^1.0.1"
+	},
+	"resolutions": {
+		"infinitton-idisplay/node-hid": "^2.1.1",
+		"electron-builder/**/app-builder-bin": "npm:@julusian/test-lib@3.5.11-app-builder-bin",
+		"osc/serialport": "^9.0.6"
 	}
 }
