@@ -293,7 +293,7 @@ graphics.prototype.drawBankImage = function (c, page, bank) {
 		if (c.png64 !== undefined) {
 
 			try {
-				var data = new Buffer(c.png64, 'base64');
+				var data = Buffer.from(c.png64, 'base64');
 				var halign = c.pngalignment.split(":",2)[0];
 				var valign = c.pngalignment.split(":",2)[1];
 				img.drawFromPNGdata(data, 0, 14, 72, 58, halign, valign);
