@@ -40,7 +40,7 @@ function image(width,height) {
 	self.lastBackgroundColor = self.rgb(0,0,0);
 
 	for (var y = 0; y < self.height; y++) {
-		var buf = new Buffer(self.width * 3); // * 3 for RGB.
+		var buf = Buffer.alloc(self.width * 3); // * 3 for RGB.
 		self.canvas.push(buf);
 	}
 
@@ -152,6 +152,8 @@ var rgb = image.prototype.rgb = function (r,g,b) {
 
 var rgbRev = image.prototype.rgbRev = function (dec) {
 	var self = this;
+
+	dec = Math.round(dec)
 
 	return  {
 		r: (dec & 0xff0000) >> 16,
@@ -618,7 +620,7 @@ image.prototype.drawPixelBuffer = function (x, y, width, height, buffer, type) {
 	}
 
 	if (type === 'base64') {
-		buffer = new Buffer(buffer, 'base64');
+		buffer = Buffer.from(buffer, 'base64');
 	}
 
 	if (buffer.length < width * height * 3) {
@@ -699,6 +701,70 @@ image.prototype.drawChar = function(x,y,char,color,fontindex,double,dummy) {
 	}
 
 	return maxX;
+};
+
+image.prototype.drawBorder = function(depth = 0, color) {
+	var self = this;
+
+	if ( depth > 0 ) {
+		if ( depth * 2 < self.width ) {
+			for (var x = 0; x < depth; x++) {
+				self.verticalLine(x, color);
+			}
+
+			for (var x = self.width - depth; x < self.width; x++) {
+				self.verticalLine(x, color);
+			}
+		}
+		else {
+			for (var x = 0; x < self.width; x++) {
+				self.verticalLine(x, color);
+			}
+		}
+
+		if ( depth * 2 < self.height ) {
+			for (var y = 0; y < depth; y++) {
+				self.horizontalLine(y, color);
+			}
+
+			for (var y = self.height - depth; y < self.height; y++) {
+				self.horizontalLine(y, color);
+			}
+		}
+		else {
+			for (var y = 0; y < self.height; y++) {
+				self.horizontalLine(y, color);
+			}
+		}
+
+		return true;
+	}
+	else {
+		return false;
+	}
+};
+
+image.prototype.drawCornerTriangle = function(depth=0, color, halign='left', valign='top') {
+	var self = this;
+
+	if ( depth > 0 && (halign == 'left' || halign == 'right') && (valign == 'top' || valign == 'bottom') ) {
+		var maxY = (depth > self.height ? self.height : depth);
+
+		for (var y = 0; y < maxY; y++) {
+			var trueY = (valign == 'bottom' ? self.height - 1 - y : y);
+
+			for (var x = 0; (x < (depth - y) && x < self.width); x++) {
+				var trueX = (halign == 'right' ? self.width - 1 - x : x);
+
+				self.pixel(trueX, trueY, color);
+			}
+		}
+
+		return true;
+	}
+	else {
+		return false;
+	}
 };
 
 image.prototype.toBase64 = function () {
