@@ -68,7 +68,7 @@ function protocol() {
 }
 
 protocol.SUPPORTED_MAJOR = 1;
-protocol.SUPPORTED_MINIOR = 0;
+protocol.SUPPORTED_MINIOR = 1;
 
 protocol.SA_HEADER_PARSER = new Parser()
 	.endianess("big")
@@ -80,6 +80,13 @@ protocol.SCMD_ADDDEVICE_PARSER = new Parser()
   .endianess("big")
   .string("serialNumber", { length: 20 })
   .uint16("deviceId");
+
+protocol.SCMD_ADDDEVICE2_PARSER = new Parser()
+.endianess("big")
+.string("serialNumber", { length: 20 })
+.uint16("deviceId")
+.uint8('keysTotal')
+.uint8('keysPerRow')
 
 protocol.SCMD_REMOVEDEVICE_PARSER = new Parser()
   .endianess("big")
@@ -114,6 +121,7 @@ protocol.SCMD_PONG = 0x02;
 protocol.SCMD_VERSION = 0x03;
 protocol.SCMD_ADDDEVICE = 0x10;
 protocol.SCMD_REMOVEDEVICE = 0x11;
+protocol.SCMD_ADDDEVICE2 = 0x12;
 protocol.SCMD_BUTTON = 0x20;
 protocol.SCMD_DRAW7272 = 0x30;
 protocol.SCMD_BRIGHTNESS = 0x31;
@@ -149,14 +157,14 @@ protocol.sendPacket = function(socket, command, data) {
 	var self = this;
 
 	if (data === undefined) {
-		data = new Buffer('');
+		data = Buffer.from('');
 	}
 
 	if (typeof data !== 'object' || !(data instanceof Buffer)) {
-		data = new Buffer(data);
+		data = Buffer.from(data);
 	}
 
-	var buffer = new Buffer(5 + data.length + 1);
+	var buffer = Buffer.alloc(5 + data.length + 1);
 
 	var header = {
 		signature: 0x5A7E,
