@@ -53,7 +53,7 @@ satelliteServer.prototype.initSocket = function (socket) {
 	var self = this;
 	debug('new connection from ' + socket.name);
 
-	socket.buffer = new Buffer('');
+	socket.buffer = Buffer.from('');
 
 	socket.on('data', function (data) {
 		socket.buffer = Buffer.concat([ socket.buffer, data ]);
@@ -63,6 +63,19 @@ satelliteServer.prototype.initSocket = function (socket) {
 
 	socket.on('error', function (e) {
 		debug('socket error:', e);
+	});
+
+	socket.on('close', function () {
+		for (var key in self.devices) {
+			if (self.devices[key].socket === socket) {
+				elgatoDM.removeDevice(self.devices[key].id);
+			  system.removeAllListeners(self.devices[key].id + '_button');
+				delete self.devices[key];
+			}
+		}
+
+		socket.removeAllListeners('data');
+		socket.removeAllListeners('close');
 	});
 };
 
@@ -84,6 +97,7 @@ satelliteServer.prototype.findPackets = function(socket) {
 
 		// Loop until it is found or we are out of buffer-data
 		self.findPackets(socket);
+		return;
 	}
 
 	if (header.length + 6 <= socket.buffer.length) {
@@ -117,6 +131,10 @@ satelliteServer.prototype.parsePacket = function(socket, header) {
 
 		case protocol.SCMD_ADDDEVICE:
 			self.newDevice(packet, socket);
+			break;
+
+		case protocol.SCMD_ADDDEVICE2:
+			self.newDevice2(packet, socket);
 			break;
 
 		case protocol.SCMD_REMOVEDEVICE:
@@ -173,7 +191,7 @@ satelliteServer.prototype.handleButton = function(packet, socket) {
 	var obj = protocol.SCMD_BUTTON_PARSER.parse(packet);
 
 	if (self.devices[obj.deviceId] !== undefined) {
-		self.system.emit(self.devices[obj.deviceId].id + '_button', obj.keyIndex, obj.state);
+		self.system.emit(self.devices[obj.deviceId].id + '_button', obj.keyIndex, obj.state == 1);
 	}
 };
 
@@ -208,41 +226,61 @@ satelliteServer.prototype.newDevice = function(packet, socket) {
 
 	var data = protocol.SCMD_ADDDEVICE_PARSER.parse(packet);
 
+	var internalId = self.newDevice_inner(data, socket)
+
+	// Rewrite packet, and send back as ack
+	data.deviceId = internalId;
+	protocol.SCMD_ADDDEVICE_PARSER.serialize(data, packet);
+	protocol.sendPacket(socket, protocol.SCMD_ADDDEVICE, packet);
+};
+
+satelliteServer.prototype.newDevice2 = function(packet, socket) {
+	var self = this;
+
+	var data = protocol.SCMD_ADDDEVICE2_PARSER.parse(packet);
+
+	var internalId = self.newDevice_inner(data, socket)
+
+	// Rewrite packet, and send back as ack
+	data.deviceId = internalId;
+	protocol.SCMD_ADDDEVICE2_PARSER.serialize(data, packet);
+	protocol.sendPacket(socket, protocol.SCMD_ADDDEVICE2, packet);
+};
+
+
+satelliteServer.prototype.newDevice_inner = function(data, socket) {
+	var self = this;
+
 	var internalId = self.findNewId();
 
 	debug('add device: ' + socket.remoteAddress, data.serialNumber);
 
 	// Use ip right now, since the pluginUUID is new on each boot and makes Companion
 	// forget all settings for the device. (page and orientation)
-	id = 'satellite-' + data.serialNumber;
+	var id = 'satellite-' + data.serialNumber;
+
+	const existing = Object.entries(self.devices).find(([internalId, dev]) => dev.id === id && dev.socket === socket)
+	if (existing) {
+		// Reuse the existing, to avoid duplicates issues
+		setImmediate(() => {
+			system.emit('elgato_ready', id);
+		})
+		return existing[0]
+	}
 
 	self.devices[internalId] = {
 		id: id,
 		socket: socket
 	};
 
-	elgatoDM.addDevice({ path: id }, 'satellite_device');
+	elgatoDM.addDevice({ path: id, keysTotal: data.keysTotal, keysPerRow: data.keysPerRow }, 'satellite_device');
 
-	// Rewrite packet, and send back as ack
-	data.deviceId = internalId;
-	protocol.SCMD_ADDDEVICE_PARSER.serialize(data, packet);
-	protocol.sendPacket(socket, protocol.SCMD_ADDDEVICE, packet);
+	setImmediate(() => {
+		// Give satellite_device reference to socket
+		system.emit(id + '_satellite_startup', socket, internalId);
+	})
 
-	// Give satellite_device reference to socket
-	system.emit(id + '_satellite_startup', socket, internalId);
-
-	socket.on('close', function () {
-		for (var key in self.devices) {
-			if (self.devices[key].socket === socket) {
-				elgatoDM.removeDevice(self.devices[key].id);
-			  system.removeAllListeners(self.devices[key].id + '_button');
-				delete self.devices[key];
-			}
-		}
-
-		socket.removeAllListeners('data');
-		socket.removeAllListeners('close');
-	});
+	return internalId
 };
 
 exports = module.exports = function (_system) {
