@@ -233,8 +233,8 @@ function telnetStream(options) {
 
 	this._options = options = options || {};
 
-	this.buffer = new Buffer('');
-	this.subbuffer = new Buffer('');
+	this.buffer = Buffer.from('');
+	this.subbuffer = Buffer.from('');
 	this.state = DATA;
 
 	debug("new telnetStreamer(", options, ")");
@@ -265,7 +265,7 @@ telnetStream.prototype._handleByte = function(byte) {
 			return;
 		}
 
-		self.buffer = Buffer.concat([ self.buffer, new Buffer([ byte ]) ]);
+		self.buffer = Buffer.concat([ self.buffer, Buffer.from([ byte ]) ]);
 	}
 
 	else if (self.state === IAC) {
@@ -295,11 +295,11 @@ telnetStream.prototype._handleByte = function(byte) {
 		if (byte === SE) {
 			self.emit('sb', self.subbuffer);
 			self.state = DATA;
-			self.subbuffer = new Buffer('');
+			self.subbuffer = Buffer.from('');
 			return;
 		}
 
-		self.subbuffer = Buffer.concat([ self.subbuffer, new Buffer([byte]) ]);
+		self.subbuffer = Buffer.concat([ self.subbuffer, Buffer.from([byte]) ]);
 	}
 
 };
@@ -308,7 +308,7 @@ telnetStream.prototype._getData = function(arguments) {
 	var self = this;
 	var buff = self.buffer;
 
-	self.buffer = new Buffer('');
+	self.buffer = Buffer.from('');
 
 	return buff;
 };
