@@ -60,7 +60,7 @@ common.prototype.handleBuffer = function(buffer) {
 	var self = this;
 
 	if (buffer.type == 'Buffer') {
-		buffer = new Buffer(buffer.data);
+		buffer = Buffer.from(buffer.data);
 	}
 
 	if (buffer === undefined || buffer.length != 15552) {
@@ -70,7 +70,7 @@ common.prototype.handleBuffer = function(buffer) {
 	}
 
 	if (self.config.rotation === -90) {
-		var buf = new Buffer(15552);
+		var buf = Buffer.alloc(15552);
 
 		for (var x = 0; x < 72; ++x) {
 			for (var y = 0; y < 72; ++y) {
@@ -81,7 +81,7 @@ common.prototype.handleBuffer = function(buffer) {
 	}
 
 	if (self.config.rotation === 180) {
-		var buf = new Buffer(15552);
+		var buf = Buffer.alloc(15552);
 
 		for (var x = 0; x < 72; ++x) {
 			for (var y = 0; y < 72; ++y) {
@@ -92,7 +92,7 @@ common.prototype.handleBuffer = function(buffer) {
 	}
 
 	if (self.config.rotation === 90) {
-		var buf = new Buffer(15552);
+		var buf = Buffer.alloc(15552);
 
 		for (var x = 0; x < 72; ++x) {
 			for (var y = 0; y < 72; ++y) {
