@@ -28,8 +28,8 @@ function action(system) {
 	self.bank_release_actions = {};
 	self.bank_status = {};
 	self.instance = {};
-	self.actions_running = [];
-	self.timers_running = [];
+	self.actions_running = new Set();
+	self.timers_running = new Map();
 
 	self.system.emit('db_get', 'bank_actions', function(res) {
 		if (res !== undefined) {
@@ -57,9 +57,39 @@ function action(system) {
 		}
 	});
 
+	function cleanup_actions_list(actions) {
+		const res = {}
+
+		for (var page in actions) {
+			res[page] = {}
+			for (var bank in actions[page]) {
+				res[page][bank] = []
+
+				if (actions[page][bank] !== undefined) {
+					for (var i = 0; i < actions[page][bank].length ; ++i) {
+						const action = actions[page][bank][i]
+						if (action && self.instance.store.db[action.instance]) {
+							res[page][bank].push(action)
+						}
+					}
+				}
+			}
+		}
+
+		return res
+	}
+
 	self.system.on('instance', function(obj) {
 		debug('got instance');
 		self.instance = obj;
+
+		// ensure all actions are valid
+		self.bank_actions = cleanup_actions_list(self.bank_actions)
+		self.bank_release_actions = cleanup_actions_list(self.bank_release_actions)
+
+		self.system.emit('db_set', 'bank_actions', self.bank_actions);
+		self.system.emit('db_set', 'bank_release_actions', self.bank_release_actions);
+		self.system.emit('db_save');
 	});
 
 	self.system.on('action_save', function() {
@@ -222,7 +252,7 @@ function action(system) {
 	});
 
 	self.system.on('action_running_get', function (page, bank, cb) {
-		cb(self.actions_running.indexOf(page + '_' + bank) !== -1);
+		cb(self.actions_running.has(`${page}_${bank}`));
 	});
 
 	// If a user wants to panic-abort all timers running
@@ -230,19 +260,49 @@ function action(system) {
 
 		debug("Aborting delayed actions");
 
-		while(self.timers_running.length > 0) {
+		for (let timer of self.timers_running.keys()) {
 			debug("clearing timer");
-			clearTimeout( self.timers_running.shift() );
+			clearTimeout(timer);
 		}
+		self.timers_running.clear()
 
-		var actions_running = self.actions_running.slice(0); //clone hack
-		self.actions_running = []; // clear the array
+		var actions_running = self.actions_running; //clone hack
+		self.actions_running = new Set(); // clear the array
 
-		for (var bid in actions_running) {
-			var a = actions_running[bid].split("_");
+		for (let bid of actions_running.keys()) {
+			const a = bid.split("_");
 			self.system.emit('graphics_bank_invalidate', a[0], a[1]);
 		}
 
+	});
+
+	// If a user wants to abort a single button actions
+	self.system.on('action_abort_bank', function(page, bank, unlatch) {
+
+		var bid = page + '_' + bank;
+		var cleared = 0;
+
+		self.actions_running.delete(bid);
+
+		self.timers_running.forEach((timerId, timer) => {
+			if (timerId === bid) {
+				if (cleared == 0) {
+					debug("Aborting button ",page,",",bank);
+				}
+				clearTimeout(timer);
+				self.timers_running.delete(timer);
+				cleared += 1;
+			}
+		})
+
+		// if requested, reset and skip up-actions
+		if (unlatch) {
+			self.system.emit('graphics_indicate_push', page, bank, false);
+		}
+
+		if (cleared > 0) {
+			self.system.emit('graphics_bank_invalidate', page, bank);
+		}
 	});
 
 	// skipNext needed for 'bank_pressed' callback
@@ -273,22 +333,27 @@ function action(system) {
 				}
 			}
 
-			let reject = false
+			let reject = false;
 			system.emit('graphics_is_pushed', page, bank, function (pushed) {
+				let isPushed = (1 == pushed? true : false);
 				// button is being pressed but not yet latched
 				// the next button-release from this device needs to be skipped
 				// because the 'release' would immediately un-latch the button
-				if (direction && (pushed != 1)) {
+				if (direction && !isPushed) {
 					skipNext[pb] = deviceid;
 				} else if (direction && pushed) {
 					// button is latched, prevent duplicate down actions
 					// the following 'release' will run the up actions
-					reject = true
+					reject = true;
+				} else if (!(direction || pushed)) {
+					// button is up, prevent duplicate up actions
+					reject = true;
 				}
 			});
 
 			if (reject) {
-				return
+				//debug("Latch button duplicate " + (direction? "down":"up") )
+				return;
 			}
 		}
 
@@ -348,26 +413,12 @@ function action(system) {
 			a.effective_delay = action_delay;
 		}
 
-		var maxtime = 0;
-		var maxidx = -1;
+		const bankId = `${page}_${bank}`
 
+		let has_delayed = false;
 		for (var n in obj[page][bank]) {
 			var a = obj[page][bank][n];
-			if (a.effective_delay !== undefined && parseInt(a.effective_delay) > maxtime) {
-				maxtime = parseInt(a.effective_delay);
-				maxidx = n;
-			}
-		}
-
-		// Start timer-indication
-		if (maxtime > 0) {
-			self.actions_running.push(page + '_' + bank);
-		}
-
-		var has_delayed = false;
-		for (var n in obj[page][bank]) {
-			var a = obj[page][bank][n];
-			var delay = parseInt(a.effective_delay === undefined ? 0 : a.effective_delay);
+			var delay = a.effective_delay === undefined ? 0 : parseInt(a.effective_delay);
 			delete a.effective_delay;
 
 			debug("Running action", a);
@@ -380,30 +431,24 @@ function action(system) {
 
 						has_delayed = true;
 
-						(function(action, delay_time, n) {
+						(function(action, delay_time) {
 							var timer = setTimeout(function() {
 								self.system.emit('action_run', action, { deviceid: deviceid, page: page, bank: bank });
 
-								// Stop timer-indication
-								if (maxtime > 0 && maxidx == n) {
-									var idx;
-									if ((idx = self.actions_running.indexOf(page + '_' + bank)) !== -1) {
-										self.actions_running.splice(idx, 1);
-										self.system.emit('graphics_bank_invalidate', page, bank);
-									}
-								}
+								self.timers_running.delete(timer);
 
-								// Remove myself from running timers
-								var idx = self.timers_running.indexOf(timer);
-								if (idx !== -1) {
-									self.timers_running.splice(idx, 1);
+								// Stop timer-indication
+								const hasAnotherTimer = Array.from(self.timers_running.values()).find(v => v === bankId)
+								if (hasAnotherTimer === undefined) {
+									self.actions_running.delete(bankId)
+									self.system.emit('graphics_bank_invalidate', page, bank);
 								}
 
 							}, delay_time);
 
-							self.timers_running.push(timer);
+							self.timers_running.set(timer, bankId);
 
-						})(a, delay, n);
+						})(a, delay);
 					}
 
 					// or is it immediate
@@ -421,6 +466,9 @@ function action(system) {
 		}
 
 		if (has_delayed) {
+			// Start timer-indication
+			self.actions_running.add(bankId);
+
 			self.system.emit('graphics_bank_invalidate', page, bank);
 		}
 	});
@@ -432,24 +480,28 @@ function action(system) {
 				if (self.instance.store.db[action.instance] !== undefined && self.instance.store.db[action.instance].enabled !== false) {
 					const definition = self.actions[`${action.instance}:${action.action}`]
 
-					// Ask instance to execute action
-					if (definition !== undefined && definition.callback !== undefined && typeof definition.callback == 'function') {
-						definition.callback(action, extras);
-					} else if (instance !== undefined && typeof instance.action == 'function') {
-						/*
-						  https://github.com/bitfocus/companion/issues/1117
-						  https://sentry.bitfocus.io/organizations/bitfocus/issues/11/?project=8
-						 
-						  the line above was "fixed" by the 'instance !== undefined' to 
-						  avoid surprise features/bugs.. 
-						  I don't have a clue what's going on here.. Help wanted. -WV
-						  Sentry tells us that 33 users got this 236 times over 11 days
-						*/
-						instance.action(action, extras);
-					} else {
-						debug('ERROR: instance does not have an action() function:', instance);
-					}
+					try {
+						// Ask instance to execute action
+						if (definition !== undefined && definition.callback !== undefined && typeof definition.callback == 'function') {
+							definition.callback(action, extras);
+						} else if (instance !== undefined && typeof instance.action == 'function') {
+							/*
+							  https://github.com/bitfocus/companion/issues/1117
+							  https://sentry.bitfocus.io/organizations/bitfocus/issues/11/?project=8
 
+							  the line above was "fixed" by the 'instance !== undefined' to
+							  avoid surprise features/bugs..
+							  I don't have a clue what's going on here.. Help wanted. -WV
+							  Sentry tells us that 33 users got this 236 times over 11 days
+							*/
+							instance.action(action, extras);
+						} else {
+							debug('ERROR: instance does not have an action() function:', instance);
+						}
+					}
+					catch(e) {
+						self.system.emit('log', 'instance('+instance.label+')', 'warn', 'Error executing action: ' + e.message);
+					}
 				}
 				else {
 					debug("trying to run action on a deleted instance.", action)
@@ -543,6 +595,11 @@ function action(system) {
 					'options': {}
 				};
 
+				if (!self.instance.store.db[act.instance]) {
+					// Action is not valid
+					return
+				}
+
 				if (self.actions[action] !== undefined) {
 					var definition = self.actions[action];
 
@@ -574,6 +631,11 @@ function action(system) {
 					'action': s[1],
 					'options': {}
 				};
+
+				if (!self.instance.store.db[act.instance]) {
+					// Action is not valid
+					return
+				}
 
 				if (self.actions[action] !== undefined) {
 					var definition = self.actions[action];
@@ -745,6 +807,18 @@ function action(system) {
 	});
 
 	self.system.on('instance_actions', function(id, actions) {
+		let newActions = {};
+		// actions[instance_id:action]
+		for (var m in self.actions) {
+			var idActionSplit = m.split(':');
+			// only add other id actions.
+			if (idActionSplit[0] != id) {
+				newActions[m] = self.actions[m];
+			}
+		}
+		// overwrite old action array with cleared one
+		self.actions = newActions;
+
 		for (var n in actions) {
 			var a = actions[n];
 			self.actions[id+':'+n] = a;
@@ -761,7 +835,7 @@ action.prototype.subscribeAction = function(action) {
 
 	console.log(action)
 
-	if (action.action !== undefined && action.instance !== undefined) {
+	if (action !== undefined && action.action !== undefined && action.instance !== undefined) {
 		const actionId = `${action.instance}:${action.action}`
 		if (self.actions[actionId] !== undefined) {
 			let definition = self.actions[actionId];
@@ -776,7 +850,7 @@ action.prototype.subscribeAction = function(action) {
 action.prototype.unsubscribeAction = function(action) {
 	var self = this;
 
-	if (action.action !== undefined && action.instance !== undefined) {
+	if (action !== undefined && action.action !== undefined && action.instance !== undefined) {
 		const actionId = `${action.instance}:${action.action}`
 		if (self.actions[actionId] !== undefined) {
 			let definition = self.actions[actionId];
