@@ -116,7 +116,7 @@ function bank(system) {
 				type: 'textinput',
 				id: 'text',
 				label: 'Text',
-				width: 7,
+				width: 6,
 				default: ''
 			},
 
@@ -142,7 +142,7 @@ function bank(system) {
 				id: 'png',
 				label: '72x58 PNG',
 				accept: 'image/png',
-				width: 2,
+				width: 3,
 				imageMinWidth: 72,
 				imageMinHeight: 58,
 				imageMaxWidth: 72,
@@ -243,7 +243,7 @@ function bank(system) {
 				if (self.config[x] === undefined) {
 					self.config[x] = {};
 					for (var y = 1; y <= global.MAX_BUTTONS; y++) {
-						if (self.config[y] === undefined) {
+						if (self.config[x][y] === undefined) {
 							self.config[x][y] = {};
 						}
 					}
@@ -340,6 +340,14 @@ function bank(system) {
 			client.emit('bank_set_png:result', 'ok');
 			system.emit('graphics_bank_invalidate', page, bank);
 		});
+
+		client.on('bank_clear_png', function(page, bank) {
+			delete self.config[page][bank].png64;
+			system.emit('bank_update', self.config);
+
+			client.emit('bank_clear_png:result');
+			system.emit('graphics_bank_invalidate', page, bank);
+		})
 
 		client.on('bank_changefield', function(page, bank, key, val) {
 			system.emit('bank_changefield', page, bank, key, val);
