@@ -131,22 +131,24 @@ function device(_system, panel) {
 		}
 	}, 1000);
 
-	system.on('device_redraw', function (id) {
+	self.on_device_redraw = function (id) {
 		if (id == self.devicepath) {
 			self.drawPage();
 		}
-	});
+	}
+	system.on('device_redraw', self.on_device_redraw);
 
 	system.on('elgato_ready',self.on_elgato_ready);
 
-	system.on('device_page_set', function (deviceid, page) {
+	self.on_device_page_set = function (deviceid, page) {
 		if (self.panel.serialnumber == deviceid) {
 			self.page = page;
 			if (self.page == 100) { self.page = 1; }
 			if (self.page == 0) { self.page = 99; }
 			self.updatePagedevice();
 		}
-	});
+	}
+	system.on('device_page_set', self.on_device_page_set);
 
 	self.lockoutDevice = function (deviceid) {
 		if (self.panel.serialnumber == deviceid) {
@@ -168,29 +170,23 @@ function device(_system, panel) {
 
 	system.on('unlockout_device', self.unlockoutDevice);
 
-	system.on('device_page_down', function (deviceid) {
+	self.on_device_page_down = function (deviceid) {
 		if (self.userconfig.page_direction_flipped === true) {
 			deviceIncreasePage(deviceid);
 		} else {
 			deviceDecreasePage(deviceid);
 		}
-	});
+	}
+	system.on('device_page_down', self.on_device_page_down);
 
-	system.on('device_page_up', function (deviceid) {
+	self.on_device_page_up = function (deviceid) {
 		if (self.userconfig.page_direction_flipped === true) {
 			deviceDecreasePage(deviceid);
 		} else {
 			deviceIncreasePage(deviceid);
 		}
-	});
-
-	system.on('device_inc_page', function (deviceid) {
-		deviceIncreasePage(deviceid);
-	});
-
-	system.on('device_dec_page', function (deviceid) {
-		deviceDecreasePage(deviceid);
-	});
+	}
+	system.on('device_page_up', self.on_device_page_up);
 
 	function deviceIncreasePage(deviceid) {
 		if (self.panel.serialnumber == deviceid) {
@@ -213,11 +209,9 @@ function device(_system, panel) {
 	}
 
 	self.on_elgato_click = function(devicepath, key, state, obj) {
-
 		if (devicepath != self.devicepath) {
 			return;
 		}
-
 		if (!self.lockedout) {
 
 			self.lastinteraction = Date.now();
@@ -415,6 +409,10 @@ device.prototype.unload = function () {
 	system.removeListener('unlockoutall', self.unlockoutAll);
 	system.removeListener('lockout_device', self.lockoutDevice);
 	system.removeListener('unlockout_device', self.unlockoutDevice);
+	system.removeListener('device_redraw', self.on_device_redraw);
+	system.removeListener('device_page_set', self.on_device_page_set);
+	system.removeListener('device_page_down', self.on_device_page_down);
+	system.removeListener('device_page_up', self.on_device_page_up);
 	self.panel.device = undefined;
 };
 
