@@ -40,12 +40,37 @@ function elgatoDM(_system) {
 
 	system.on('devices_list_get', function (cb) {
 		var ary = [];
+		var keys = Object.keys(instances);
 
-		for (var id in instances) {
-			ary.push({ id: instances[id].id, serialnumber: instances[id].serialnumber, type: instances[id].type, config: instances[id].config });
+		if (keys[0] != undefined) {
+			ary.push({ id: instances[keys[0]].id, serialnumber: instances[keys[0]].serialnumber, type: instances[keys[0]].type, config: instances[keys[0]].config });
 		}
 
+		var item = 1;
+		while(item < keys.length) {
+
+			if (instances[keys[item]].type === 'emulator') {
+				ary.splice(0, 0, { id: instances[keys[item]].id, serialnumber: instances[keys[item]].serialnumber, type: instances[keys[item]].type, config: instances[keys[item]].config });
+			} else {
+				var i = 0;
+				while(  ary[i] !== undefined && (
+					ary[i].id == 'emulator' ||
+					ary[i].type < instances[keys[item]].type ||
+					(ary[i].type == instances[keys[item]].type && ary[i].serialnumber < instances[keys[item]].serialnumber))
+					) {
+					i++;
+				}
+				ary.splice(i, 0, { id: instances[keys[item]].id, serialnumber: instances[keys[item]].serialnumber, type: instances[keys[item]].type, config: instances[keys[item]].config });
+			}
+			item++;
+
+		}
+		
 		cb(ary);
+	});
+
+	system.on('devices_reenumerate', function() {
+		self.refreshDevices();
 	});
 
 	system.emit('io_get', function (io) {
@@ -54,10 +79,6 @@ function elgatoDM(_system) {
 		system.on('io_connect', function (socket) {
 			socket.on('devices_list_get', function () {
 				self.updateDevicesList(socket);
-			});
-
-			system.on('devices_reenumerate', function() {
-				self.refreshDevices();
 			});
 
 			socket.on('devices_reenumerate', function () {
@@ -101,9 +122,30 @@ function elgatoDM(_system) {
 elgatoDM.prototype.updateDevicesList = function (socket) {
 	var self = this;
 	var ary = [];
+	var keys = Object.keys(instances);
 
-	for (var id in instances) {
-		ary.push({ id: instances[id].id, serialnumber: instances[id].serialnumber, type: instances[id].type, config: instances[id].config });
+  if (keys[0] != undefined) {
+	  ary.push({ id: instances[keys[0]].id, serialnumber: instances[keys[0]].serialnumber, type: instances[keys[0]].type, config: instances[keys[0]].config });
+  }
+
+	var item = 1;
+	while(item < keys.length) {
+
+		if (instances[keys[item]].type === 'emulator') {
+			ary.splice(0, 0, { id: instances[keys[item]].id, serialnumber: instances[keys[item]].serialnumber, type: instances[keys[item]].type, config: instances[keys[item]].config });
+		} else {
+			var i = 0;
+			while(  ary[i] !== undefined && (
+				ary[i].id == 'emulator' ||
+				ary[i].type < instances[keys[item]].type ||
+				(ary[i].type == instances[keys[item]].type && ary[i].serialnumber < instances[keys[item]].serialnumber))
+				) {
+				i++;
+			}
+			ary.splice(i, 0, { id: instances[keys[item]].id, serialnumber: instances[keys[item]].serialnumber, type: instances[keys[item]].type, config: instances[keys[item]].config });
+		}
+		item++;
+
 	}
 
 	system.emit('devices_list', ary);
@@ -122,7 +164,6 @@ elgatoDM.prototype.refreshDevices = function (cb) {
 	function scanUSB() {
 		debug("USB: checking devices (blocking call)");
 		var devices = HID.devices();
-
 		for (var i = 0; i < devices.length; ++i) {
 			(function (device) {
 				if (!ignoreStreamDeck && device.vendorId === 0x0fd9 && device.productId === 0x0060 && instances[device.path] === undefined) {
@@ -142,6 +183,9 @@ elgatoDM.prototype.refreshDevices = function (cb) {
 				}
 				else if (device.vendorId === 0xffff && device.productId === 0x1f41 && instances[device.path] === undefined) {
 					self.addDevice(device, 'infinitton');
+				}
+				else if (device.vendorId === 1523 && device.interface === 0) {
+					self.addDevice(device, 'xkeys');
 				}
 
 			})(devices[i]);
@@ -179,15 +223,19 @@ elgatoDM.prototype.refreshDevices = function (cb) {
 			scanUSB();
 		}
 	} catch (e) {
-		// scan for all usb devices anyways
-		scanUSB();
+		try {
+			// scan for all usb devices anyways
+			scanUSB();
+		} catch (e) {
+			debug('USB: scan failed')
+		}
 	}
 };
 
 elgatoDM.prototype.addDevice = function (device, type) {
 	var self = this;
 	debug('add device ' + device.path);
-
+	
 	if (instances[device.path] !== undefined) {
 		try {
 			instances[device.path].quit();
@@ -209,7 +257,7 @@ elgatoDM.prototype.addDevice = function (device, type) {
 
 		self.updateDevicesList();
 	} else if (type === 'satellite_device') {
-		instances[device.path] = new satelliteDevice(system, device.path);
+		instances[device.path] = new satelliteDevice(system, device.path, device);
 		instances[device.path].deviceHandler = new deviceHandler(system, instances[device.path]);
 
 		self.updateDevicesList();
