@@ -54,6 +54,35 @@ function feedback(system) {
 		}
 	});
 
+	self.system.on('instance', function(obj) {
+		self.instance = obj
+		debug('got instance');
+
+		// ensure all feedbacks are valid
+		const res = {}
+		for (var page in self.feedbacks) {
+			res[page] = {}
+			for (var bank in self.feedbacks[page]) {
+				res[page][bank] = []
+
+				// Iterate through feedbacks on this bank
+				if (self.feedbacks[page][bank] !== undefined) {
+					for (var i in self.feedbacks[page][bank]) {
+						var feedback = self.feedbacks[page][bank][i];
+
+						if (feedback && self.instance.store.db[feedback.instance_id]) {
+							res[page][bank].push(feedback)
+						}
+					}
+				}
+			}
+		}
+		self.feedbacks = res
+
+		self.system.emit('db_set', 'feedbacks', self.feedbacks);
+		self.system.emit('db_save');
+	});
+
 	system.on('feedback_instance_definitions_set', function (instance, feedbacks) {
 		self.feedback_definitions[instance.id] = feedbacks;
 		io.emit('feedback_get_definitions:result', self.feedback_definitions);
@@ -166,15 +195,20 @@ function feedback(system) {
 									definition = self.feedback_definitions[instance.id][feedback.type];
 								}
 
-								// Ask instance to check bank for custom styling
-								if (definition !== undefined && definition.callback !== undefined && typeof definition.callback == 'function') {
-									var result = definition.callback(feedback, bank_obj);
-									self.setStyle(page, bank, i, result);
-								} else if (typeof instance.feedback == 'function') {
-									var result = instance.feedback(feedback, bank_obj);
-									self.setStyle(page, bank, i, result);
-								} else {
-									debug('ERROR: instance ' + instance.label + ' does not have a feedback() function');
+								try {
+									// Ask instance to check bank for custom styling
+									if (definition !== undefined && definition.callback !== undefined && typeof definition.callback == 'function') {
+										var result = definition.callback(feedback, bank_obj);
+										self.setStyle(page, bank, i, result);
+									} else if (typeof instance.feedback == 'function') {
+										var result = instance.feedback(feedback, bank_obj);
+										self.setStyle(page, bank, i, result);
+									} else {
+										debug('ERROR: instance ' + instance.label + ' does not have a feedback() function');
+									}
+								}
+								catch(e) {
+									self.system.emit('log', 'instance('+instance.label+')', 'warn', 'Error checking feedback: ' + e.message);
 								}
 							});
 						}
@@ -261,6 +295,11 @@ function feedback(system) {
 				'type': s[1],
 				'instance_id': s[0],
 				'options': {}
+			}
+
+			if (!self.instance.store.db[fb.instance_id]) {
+				// Feedback is not valid
+				return
 			}
 
 			if (self.feedback_definitions[s[0]] !== undefined && self.feedback_definitions[s[0]][s[1]] !== undefined) {
