@@ -31,6 +31,7 @@ function instance(system) {
 	self.modules_manufacturer = {};
 	self.modules_category = {};
 	self.modules_name = {};
+	self.modules_main = {};
 	self.package_info = {};
 	self.status = {};
 
@@ -43,8 +44,6 @@ function instance(system) {
 	self.numWarn = 0;
 	self.numOk = 0;
 	
-	system.emit('instance', self);
-
 	system.on('instance_get_package_info', function(cb) {
 		debug("getting instance_get_package_info");
 		cb(self.package_info);
@@ -83,7 +82,15 @@ function instance(system) {
 		self.system.emit('log', 'instance(' + id + ')', 'info', 'instance deleted');
 
 		if (self.active[id] !== undefined) {
-			self.active[id].destroy();
+
+			if (self.active[id].destroy !== undefined && typeof self.active[id].destroy == 'function') {
+				try {
+					self.active[id].destroy();
+				} catch (e) {
+					self.system.emit('log', 'instance('+id+')', 'debug', 'Error while deleting instance: ' + e.message);
+				}
+			}
+
 			delete self.active[id];
 		}
 		delete self.status[id];
@@ -108,7 +115,7 @@ function instance(system) {
 
 			for (var id in self.active) {
 
-				if (self.active[id] !== undefined) {
+				if (self.active[id] !== undefined && self.active[id].destroy !== undefined && typeof self.active[id].destroy == 'function') {
 					try {
 						self.active[id].destroy();
 					} catch (e) {
@@ -136,7 +143,7 @@ function instance(system) {
 			}
 		}
 	});
-
+	
 	system.on('instance_init', function() {
 
 		debug('instance_init', self.store.db);
@@ -153,15 +160,20 @@ function instance(system) {
 
 				var packagefile = fs.readFileSync(path +'/'+ folder + '/package.json');
 				var moduleconfig = JSON.parse(packagefile);
-				var mod = require(path + '/' + folder + '/' + moduleconfig.main);
+				var main = path + '/' + folder + '/' + moduleconfig.main;
 				if (fs.existsSync(path +'/'+ folder + '/HELP.md')) {
 					moduleconfig.help = true;
 				} else {
 					moduleconfig.help = false;
 				}
 
+				if (process.env.LOAD_ALL_MODULES !== undefined) {
+					var mod = require(path + '/' + folder + '/' + moduleconfig.main);
+					self.modules[folder] = mod;
+				}
+
 				self.store.module.push(moduleconfig);
-				self.modules[folder] = mod;
+				self.modules_main[folder] = main;
 				self.package_info[folder] = moduleconfig;
 
 				// Generate keywords list
@@ -175,19 +187,19 @@ function instance(system) {
 					}
 				}
 				else {
-					console.log(folder, "- uh, no keywords?")
+					console.log(folder, "- uh, no keywords?");
 					process.exit();
 				}
 
 				// Generate manufacturer list
-				if (moduleconfig.manufacturer !== undefined && moduleconfig.manufacturer != 'Bitfocus') {
+				if (moduleconfig.manufacturer !== undefined && moduleconfig.name !== 'bitfocus-companion') {
 					if (self.modules_manufacturer[moduleconfig.manufacturer] === undefined) {
 						self.modules_manufacturer[moduleconfig.manufacturer] = [];
 					}
 					self.modules_manufacturer[moduleconfig.manufacturer].push(folder);
 				}
-				else if (moduleconfig.manufacturer != 'Bitfocus') {
-					console.log(folder, "- uh, no manufacturer?")
+				else if (moduleconfig.manufacturer !== 'Bitfocus') {
+					console.log(folder, "- uh, no manufacturer?");
 					process.exit();
 				}
 
@@ -200,7 +212,7 @@ function instance(system) {
 				}
 
 				// Generate label
-				if (moduleconfig.name !== undefined && moduleconfig.manufacturer != 'Bitfocus') {
+				if (moduleconfig.name !== undefined && moduleconfig.name !== 'bitfocus-companion') {
 					if (typeof moduleconfig.product == 'string') {
 						moduleconfig.label = moduleconfig.manufacturer + ":" + moduleconfig.product;
 					}
@@ -252,7 +264,7 @@ function instance(system) {
 					config.instance_type = mod_redirects[config.instance_type];
 				}
 
-				if (self.modules[config.instance_type] !== undefined) {
+				if (self.checkModuleLoaded(config.instance_type)) {
 
 					if (config.enabled === false) {
 						debug("Won't load disabled module " + id + " (" + config.instance_type + ")");
@@ -270,7 +282,12 @@ function instance(system) {
 
 					if (typeof self.active[id]._init == 'function') {
 						debug("Running _init of " + id);
-						self.active[id]._init();
+						try {
+							self.active[id]._init();
+						}
+						catch(e) {
+							self.system.emit('log', 'instance('+self.active[id].label+')', 'warn', 'Error initalizing module: ' + e.message);
+						}
 					}
 				}
 
@@ -304,6 +321,8 @@ function instance(system) {
 		}, 2000);
 	})
 
+	system.emit('instance', self);
+
 	system.emit('io_get', function(io) {
 		self.io = io;
 
@@ -319,47 +338,50 @@ function instance(system) {
 	system.on('instance_add', function (data, cb) {
 		var module = data.type;
 		var product = data.product;
-		var mod = self.modules[module];
-		var id = shortid.generate();
 
-		self.store.db[id] = {};
+		if (self.checkModuleLoaded(module)) {
+			var mod = self.modules[module];
+			var id = shortid.generate();
 
-		self.system.emit('log', 'instance('+id+')', 'info', 'instance add ' + module + ' ' + product);
+			self.store.db[id] = {};
 
-		self.store.db[id].instance_type = module;
-		self.store.db[id].product = product;
+			self.system.emit('log', 'instance('+id+')', 'info', 'instance add ' + module + ' ' + product);
 
-		var label = self.package_info[module].shortname;
-		var i = 1;
-		var freename = false;
+			self.store.db[id].instance_type = module;
+			self.store.db[id].product = product;
 
-		while (!freename) {
-			freename = true;
-			for (var key in self.store.db) {
-				if (self.store.db[key].label == label) {
-					i++;
-					label = self.package_info[module].shortname + i;
-					freename = false;
-					break;
+			var label = self.package_info[module].shortname;
+			var i = 1;
+			var freename = false;
+
+			while (!freename) {
+				freename = true;
+				for (var key in self.store.db) {
+					if (self.store.db[key].label == label) {
+						i++;
+						label = self.package_info[module].shortname + i;
+						freename = false;
+						break;
+					}
 				}
 			}
-		}
 
-		self.store.db[id].label = label;
-		if (self.active[id] !== undefined) {
-			self.active[id].label = self.store.db[id].label;
-		}
+			self.store.db[id].label = label;
+			if (self.active[id] !== undefined) {
+				self.active[id].label = self.store.db[id].label;
+			}
 
-		self.activate_module(id, mod);
+			self.activate_module(id, mod);
 
-		self.io.emit('instance_add:result', id, self.store.db);
-		debug('instance_add', id);
-		self.system.emit('instance_save');
-		self.system.emit('actions_update')
-		self.system.emit('feedback_update')
+			self.io.emit('instance_add:result', id, self.store.db);
+			debug('instance_add', id);
+			self.system.emit('instance_save');
+			self.system.emit('actions_update')
+			self.system.emit('feedback_update')
 
-		if (typeof cb == 'function') {
-			cb(id, self.store.db[id]);
+			if (typeof cb == 'function') {
+				cb(id, self.store.db[id]);
+			}
 		}
 	});
 
@@ -385,7 +407,12 @@ function instance(system) {
 
 		if (!internal && self.active[id] !== undefined) {
 			if (typeof self.active[id].updateConfig == 'function') {
-				self.active[id].updateConfig(Object.assign({}, config));
+				try {
+					self.active[id].updateConfig(Object.assign({}, config));
+				}
+				catch(e) {
+					self.system.emit('log', 'instance('+self.active[id].label+')', 'warn', 'Error updating instance configuration: ' + e.message);
+				}
 			}
 		}
 
@@ -403,7 +430,7 @@ function instance(system) {
 		}
 		
 		if (state === false) {
-			if (self.active[id] !== undefined) {
+			if (self.active[id] !== undefined && self.active[id].destroy !== undefined && typeof self.active[id].destroy == 'function') {
 				try {
 					self.active[id].destroy();
 				} catch (e) {
@@ -429,7 +456,7 @@ instance.prototype.activate_module = function(id,modin) {
 	if (modin !== undefined) {
 		mod = modin;
 	}
-	else {
+	else if (self.checkModuleLoaded(self.store.db[id].instance_type)) {
 		mod = self.modules[self.store.db[id].instance_type];
 	}
 
@@ -446,11 +473,16 @@ instance.prototype.activate_module = function(id,modin) {
 		}
 
 		if (typeof self.active[id]._init == 'function') {
-			self.active[id]._init();
+			try {
+				self.active[id]._init();
+			}
+			catch(e) {
+				self.system.emit('log', 'instance('+self.active[id].label+')', 'error', 'Error initalizing module: ' + e.message);
+			}
 		}
 
 	} catch(e) {
-		self.system.emit('log', 'instance('+id+')', 'error', 'module failed');
+		self.system.emit('log', 'instance('+id+')', 'error', 'Instance failed to launch: ' + e.message);
 		debug("INSTANCE ADD EXCEPTION:", e);
 	}
 
@@ -484,6 +516,28 @@ instance.prototype.calculateErrors = function (inst) {
 	
 };
 
+instance.prototype.checkModuleLoaded = function(instance_type) {
+	var self = this;
+	var out = false;
+
+	if (self.modules[instance_type] === undefined && self.modules_main[instance_type] !== undefined) {
+		try {
+			var mod = require(self.modules_main[instance_type]);
+			self.modules[instance_type] = mod;
+			out = true;
+		}
+		catch(e) {
+			debug("Error loading module " + instance_type, e);
+			self.system.emit('log', 'module('+instance_type+')', 'error', 'Error loading module: '+ e);
+		}
+	}
+	else if (self.modules[instance_type] !== undefined) {
+		out = true;
+	}
+
+	return out;
+};
+
 instance.prototype.connect = function (client) {
 	var self = this;
 
@@ -501,20 +555,21 @@ instance.prototype.connect = function (client) {
 			return;
 		}
 
-		var res = self.active[id].config_fields();
-
-		res.unshift({
-			type: 'textinput',
-			id: 'label',
-			label: 'Label',
-			width: 12
-		});
+		const configFields = [
+			{
+				type: 'textinput',
+				id: 'label',
+				label: 'Label',
+				width: 12
+			},
+			...(self.active[id].config_fields() || []),
+		]
 
 		client.emit(
 			'instance_edit:result',
 			id,
 			self.store,
-			res,
+			configFields,
 			self.store.db[id]
 		);
 
@@ -565,7 +620,7 @@ instance.prototype.connect = function (client) {
 	});
 
 	client.on('instance_get_help', function (module) {
-		if (self.modules[module] !== undefined) {
+		if (self.modules_main[module] !== undefined) {
 			var path = require('app-root-path') + '/lib/module';
 
 			if (fs.existsSync(path +'/'+ module + '/HELP.md')) {
