@@ -48,7 +48,7 @@ function instance(system, id, config) {
 	self.defineConst('REGEX_PORT',          '/^([1-9]|[1-8][0-9]|9[0-9]|[1-8][0-9]{2}|9[0-8][0-9]|99[0-9]|[1-8][0-9]{3}|9[0-8][0-9]{2}|99[0-8][0-9]|999[0-9]|[1-5][0-9]{4}|6[0-4][0-9]{3}|65[0-4][0-9]{2}|655[0-2][0-9]|6553[0-4])$/');
 	self.defineConst('REGEX_PERCENT',       '/^(100|[0-9]|[0-9][0-9])$/');
 	self.defineConst('REGEX_FLOAT',         '/^([0-9]*\\.)?[0-9]+$/');
-	self.defineConst('REGEX_FLOAT_OR_INT',  '/^([0-9]+)(\\.[0-9+])?$/');
+	self.defineConst('REGEX_FLOAT_OR_INT',  '/^([0-9]+)(\\.[0-9]+)?$/');
 	self.defineConst('REGEX_SIGNED_FLOAT',  '/^[+-]?([0-9]*\\.)?[0-9]+$/');
 	self.defineConst('REGEX_NUMBER',        '/^\\d+$/');
 	self.defineConst('REGEX_SOMETHING',     '/^.+$/');
@@ -152,6 +152,11 @@ instance.prototype.upgradeConfig = function () {
 		}
 		self.config._configIdx = i;
 
+		for (const action of [...actions, release_actions]) {
+			action.instance = self.id
+			action.label = `${self.id}:${action.action}`
+		}
+
 		// If anything was changed, update system and db
 		if (result) {
 			self.system.emit('config_save');
@@ -215,6 +220,12 @@ instance.prototype.getVariable = function(variable, cb) {
 
 	self.system.emit('variable_get', self.label, variable, cb);
 };
+
+instance.prototype.parseVariables = function(string, cb) {
+	var self = this;
+
+	self.system.emit('variable_parse', string, cb);
+}
 
 instance.prototype.setFeedbackDefinitions = function(feedbacks) {
 	var self = this;
