@@ -37,64 +37,69 @@ function osc(system) {
 	});
 
 	self.listener.on("message", function (message) {
-		var a = message.address.split("/");
-		if (message.address.match(/^\/press\/bank\/\d+\/\d+$/)) {
+		try {	
+			var a = message.address.split("/");
+			if (message.address.match(/^\/press\/bank\/\d+\/\d+$/)) {
 
-			if (message.args.length == 0) {
+				if (message.args.length == 0) {
 
-				debug("Got /press/bank/ (trigger)",parseInt(a[3]),"button",parseInt(a[4]));
-				system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true);
-
-				setTimeout(function (){
-					debug("Auto releasing /press/bank/ (trigger)",parseInt(a[3]),"button",parseInt(a[4]));
-					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false);
-				}, 20);
-
-			} else {
-
-				if (message.args[0].type == 'i' && message.args[0].value == '1') {
-					debug("Got /press/bank/ (press) for bank", parseInt(a[3]), "button", parseInt(a[4]));
+					debug("Got /press/bank/ (trigger)",parseInt(a[3]),"button",parseInt(a[4]));
 					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true);
-				}
-				else if (message.args[0].type == 'i' && message.args[0].value == '0') {
-					debug("Got /press/bank/ (release) for bank", parseInt(a[3]), "button", parseInt(a[4]));
-					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false);
-				}
 
-			}
-		} else if (message.address.match(/^\/style\/bgcolor\/\d+\/\d+$/)) {
-			if (message.args.length > 2) {
-				var r = message.args[0].value;
-				var g = message.args[1].value;
-				var b = message.args[2].value;
-				if (typeof r === "number" && typeof g === "number" && typeof b === "number") {
-					var col = rgb(r, g, b);
-					debug("Got /style/bgcolor", parseInt(a[3]), "button", parseInt(a[4]))
-					system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'bgcolor', col);
-					system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
+					setTimeout(function (){
+						debug("Auto releasing /press/bank/ (trigger)",parseInt(a[3]),"button",parseInt(a[4]));
+						system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false);
+					}, 20);
+
+				} else {
+
+					if (message.args[0].type == 'i' && message.args[0].value == '1') {
+						debug("Got /press/bank/ (press) for bank", parseInt(a[3]), "button", parseInt(a[4]));
+						system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true);
+					}
+					else if (message.args[0].type == 'i' && message.args[0].value == '0') {
+						debug("Got /press/bank/ (release) for bank", parseInt(a[3]), "button", parseInt(a[4]));
+						system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false);
+					}
+
+				}
+			} else if (message.address.match(/^\/style\/bgcolor\/\d+\/\d+$/)) {
+				if (message.args.length > 2) {
+					var r = message.args[0].value;
+					var g = message.args[1].value;
+					var b = message.args[2].value;
+					if (typeof r === "number" && typeof g === "number" && typeof b === "number") {
+						var col = rgb(r, g, b);
+						debug("Got /style/bgcolor", parseInt(a[3]), "button", parseInt(a[4]))
+						system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'bgcolor', col);
+						system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
+					}
+				}
+			} else if (message.address.match(/^\/style\/color\/\d+\/\d+$/)) {
+				if (message.args.length > 2) {
+					var r = message.args[0].value;
+					var g = message.args[1].value;
+					var b = message.args[2].value;
+					if (typeof r === "number" && typeof g === "number" && typeof b === "number") {
+						var col = rgb(r, g, b);
+						debug("Got /style/color", parseInt(a[3]), "button", parseInt(a[4]))
+						system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'color', col);
+						system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
+					}
+				}
+			} else if (message.address.match(/^\/style\/text\/\d+\/\d+$/)) {
+				if (message.args.length > 0) {
+					var text = message.args[0].value;
+					if (typeof text === "string") {
+						debug("Got /style/text", parseInt(a[3]), "button", parseInt(a[4]))
+						system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'text', text);
+						system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
+					}
 				}
 			}
-		} else if (message.address.match(/^\/style\/color\/\d+\/\d+$/)) {
-			if (message.args.length > 2) {
-				var r = message.args[0].value;
-				var g = message.args[1].value;
-				var b = message.args[2].value;
-				if (typeof r === "number" && typeof g === "number" && typeof b === "number") {
-					var col = rgb(r, g, b);
-					debug("Got /style/color", parseInt(a[3]), "button", parseInt(a[4]))
-					system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'color', col);
-					system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
-				}
-			}
-		} else if (message.address.match(/^\/style\/text\/\d+\/\d+$/)) {
-			if (message.args.length > 0) {
-				var text = message.args[0].value;
-				if (typeof text === "string") {
-					debug("Got /style/text", parseInt(a[3]), "button", parseInt(a[4]))
-					system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'text', text);
-					system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
-				}
-			}
+		}
+		catch (error) {
+			system.emit('log', 'osc', 'warn', 'OSC Error: ' + error);
 		}
 	});
 
