@@ -17,26 +17,21 @@
 var EventEmitter = require('events').EventEmitter;
 var util         = require('util');
 var debug        = require('debug')('lib/satellite_device');
-var path         = require('path');
 var protocol     = require('./satellite_protocol');
 var Image        = require('./image');
 
-var system;
-var express;
-var io;
-
-function satellite(_system, devicepath) {
+function satellite(system, devicepath, deviceInfo) {
 	var self = this;
 
-	self.system = system = _system;
+	self.system = system;
 	EventEmitter.call(self);
 	self.buttonState = [];
 
 	self.type = 'Satellite device';
 	self.serialnumber = devicepath;
 	self.id = devicepath;
-	self.keysPerRow = 8;
-	self.keysTotal = 32;
+	self.keysPerRow = deviceInfo.keysPerRow ? deviceInfo.keysPerRow : 8;
+	self.keysTotal = deviceInfo.keysTotal ? deviceInfo.keysTotal :32;
 
 	debug('Adding Satellite device');
 
@@ -61,7 +56,7 @@ function satellite(_system, devicepath) {
 		self.doButton(key, state);
 	});
 
-	for (var button = 0; button < 15; button++) {
+	for (var button = 0; button < global.MAX_BUTTONS; button++) {
 		self.buttonState[button] = {
 			pressed: false
 		};
@@ -81,7 +76,8 @@ satellite.prototype.begin = function() {
 };
 
 satellite.prototype.quit = function () {
-	self.system.removeAllListeners(this.devicepath + '_button');
+	var self = this;
+	self.system.removeAllListeners(self.devicepath + '_button');
 };
 
 satellite.prototype.draw = function(key, buffer) {
@@ -92,9 +88,7 @@ satellite.prototype.draw = function(key, buffer) {
 		return false;
 	}
 
-	// TODO: Fix
-	var hack = { log: self.debug, config: self._config };
-	buffer = self.handleBuffer.call(hack, buffer);
+	buffer = self.handleBuffer(buffer)
 
 	self.fillImage(key, buffer);
 
@@ -104,31 +98,23 @@ satellite.prototype.draw = function(key, buffer) {
 satellite.prototype.doButton = function(key, state) {
 	var self = this;
 
-	self.buttonState[key].pressed = state;
-	self.system.emit('elgato_click', self.devicepath, key, state, self.buttonState);
+	const keyIndex2 = this.toGlobalKey(key)
+
+	if (self.buttonState[keyIndex2]) {
+		self.buttonState[keyIndex2].pressed = state;
+		self.system.emit('elgato_click', self.devicepath, keyIndex2, state, self.buttonState);
+	}
 };
 
 satellite.prototype.clearDeck = function() {
 	var self = this;
 	debug('elgato.prototype.clearDeck()')
-	for (var x = 0; x < 15; x++) {
+	for (var x = 0; x < self.keysTotal; x++) {
 		self.clearKey(x);
 	}
 }
 
 /* elgato-streamdeck functions */
-
-satellite.prototype.getConfig = function (cb) {
-	var self = this;
-
-	debug('getConfig');
-
-	if (typeof cb == 'function') {
-		cb(self._config);
-	}
-
-	return self._config;
-};
 
 satellite.prototype.setConfig = function (config) {
 	var self = this;
@@ -162,6 +148,9 @@ satellite.prototype.setConfig = function (config) {
 satellite.prototype.fillImage = function (keyIndex, imageBuffer) {
   var self = this;
 
+  const keyIndex2 = self.toDeviceKey(keyIndex)
+  if (keyIndex2 === -1) return
+
   if (imageBuffer.length !== 15552) {
     throw new RangeError('Expected image buffer of length 15552, got length ' + imageBuffer.length);
   }
@@ -169,7 +158,7 @@ satellite.prototype.fillImage = function (keyIndex, imageBuffer) {
 	if (self.socket !== undefined) {
 		var buf = protocol.SCMD_DRAW7272_PARSER.serialize({
 			deviceId: self.deviceId,
-			keyIndex: keyIndex,
+			keyIndex: keyIndex2,
 			   image: imageBuffer
 		});
 
@@ -195,7 +184,7 @@ satellite.prototype.clearKey = function(keyIndex) {
   var self = this;
 
 	if (self.socket !== undefined) {
-		console.log("Key index: ", keyIndex);
+		debug("Key index: " + keyIndex);
 		var buf = protocol.SCMD_DRAW7272_PARSER.serialize({ deviceId: self.deviceId, keyIndex: keyIndex, image: self.clearImage });
 		protocol.sendPacket(self.socket, protocol.SCMD_DRAW7272, buf);
 	} else {
@@ -206,7 +195,7 @@ satellite.prototype.clearKey = function(keyIndex) {
 satellite.prototype.clearAllKeys = function() {
   var self = this;
 
-  for (var i = 0; i < 15; ++i) {
+  for (var i = 0; i < self.keysTotal; ++i) {
 
 		if (self.socket !== undefined) {
 			var buf = protocol.SCMD_DRAW7272_PARSER.serialize({ deviceId: self.deviceId, keyIndex: i, image: self.clearImage });
@@ -220,7 +209,18 @@ satellite.prototype.clearAllKeys = function() {
 
 // Steal rotation code from usb/common
 var common = require('./usb/common');
-satellite.prototype.handleBuffer = common.prototype.handleBuffer;
+util.inherits(satellite, common);
 
+satellite.prototype.getConfig = function (cb) {
+	var self = this;
+
+	debug('getConfig');
+
+	if (typeof cb == 'function') {
+		cb(self._config);
+	}
+
+	return self._config;
+};
 
 module.exports = satellite;
