@@ -32,7 +32,7 @@ express.use(function(req, res, next) {
 	next();
 });
 
-express.use(Express.static(__dirname + '/../public', {
+express.use(Express.static(path.join(__dirname, '/../public'), {
 	dotfiles: 'ignore',
 	etag: true,
 	extensions: [
@@ -57,6 +57,10 @@ express.use('/js/font-awesome', Express.static(require('app-root-path') + '/node
 express.use('/js/moment', Express.static(require('app-root-path') + '/node_modules/moment'));
 express.use('/select2', Express.static(require('app-root-path') + '/node_modules/select2/dist'));
 express.use('/tributejs', Express.static(require('app-root-path') + '/node_modules/tributejs/dist'));
+
+if (process.env.DEVELOPER !== undefined) {
+	express.use('/js/dist', Express.static(require('app-root-path') + '/public/js/babel'));
+}
 
 require('app-root-path') + '';
 
