@@ -79,7 +79,7 @@ function server_tcp(system) {
 		self.server.listen(51234);
 		system.emit('log', 'TCP Server', 'info', 'Listening for TCP commands on port 51234');
 	} catch(e) {
-		system.emit('log', 'TCP Server', 'error', 'TCP Server', 'Couldnt bind to TCP port 51234');
+		system.emit('log', 'TCP Server', 'error', 'Couldn\'t bind to TCP port 51234');
 	}
 
 	return self;
