@@ -48,7 +48,7 @@ function server_udp(system) {
 		system.emit('log', 'UDP Server', 'info', 'Listening for UDP commands on port 51235');
 		server.bind(51235, '0.0.0.0');
 	} catch(e) {
-		system.emit('log', 'UDP Server', 'error', 'UDP Server', 'Couldnt bind to UDP port 51235');
+		system.emit('log', 'UDP Server', 'error', 'Couldn\'t bind to UDP port 51235');
 	}
 
 }
