@@ -144,22 +144,30 @@ tcp.prototype.write = tcp.prototype.send = function(message, cb) {
 	if (self.connected) {
 		debug('sending ' + (message !== undefined ? message.length : 'undefined') + ' bytes to', self.host, self.port)
 
-		self.socket.write(message, function(error){
-			if (error) {
-				new_status(self, STATUS_ERROR, error.message);
-				self.emit.apply(self, ['error'].concat(Array.from(arguments)));
+		try {
+			self.socket.write(message, function(error){
+				if (error) {
+					new_status(self, STATUS_ERROR, error.message);
+					self.emit.apply(self, ['error'].concat(Array.from(arguments)));
+
+					if (typeof cb == 'function') {
+						cb(error);
+					}
+					return;
+				}
+
 
 				if (typeof cb == 'function') {
-					cb(error);
+					cb();
 				}
-				return;
-			}
-
-
-			if (typeof cb == 'function') {
-				cb();
-			}
-		});
+			});
+		} catch (error) {
+			// Unhandeled socket error
+			new_status(self, STATUS_ERROR, error.message);
+			self.emit.apply(self, ['error'].concat(Array.from(arguments)));
+			self.connected = false;
+			cb(error);
+		}
 
 		return true;
 	} else {
