@@ -25,8 +25,9 @@ var drivers = {
 	'elgato-v2': require('./elgato'),
 	'elgato-mini': require('./elgato-mini'),
 	'elgato-xl': require('./elgato-xl'),
-	'infinitton': require('./infinitton')
-};
+	'infinitton': require('./infinitton'),
+	'xkeys': require('./xkeys')
+}
 
 global.MAX_BUTTONS = parseInt(process.env.MAX_BUTTONS);
 global.MAX_BUTTONS_PER_ROW = parseInt(process.env.MAX_BUTTONS_PER_ROW);
